aura-backend/requirements.txt
pydantic
google-generativeai
tavily-python
appwrite
